Add visual cursor functionality
This commit implements a cursor of adequate size that follows the end
of the string while characters are getting typed and deleted. It can
also flicker when in standby and get permanently drawn while writing.
Currently cannot move it using arrow keys.
The text's font was also changed so that all characters have the same
width.
gapbuffer.cpp and gapbuffer.h were added in preparation for the
implementation of the Gap Buffer data structure that will be used for
efficient text manipulation.