Develop text functionality
This commit further develops the handling of text within the editor,
both on display and code wise. Now the characters are written into a
Gap Buffer, thus allowing somewhat more efficient writing inside the
text. Scrolling is also now possible (but only left and right), and so
the editing implementation on a single line of text is mostly finished.